vscode: project-updater: fix multiroot try-update-template

+++ scripts/project-updater.xsh
@@ -504,6 +504,7 @@ with open(tasks_path, "r") as f:
 
 # Get merge instructions
 merge_config = _template_metadata.get("mergeCommon", {})
+has_dockerfile = _template_metadata.get("hasDockerfile", True)
 task_labels_to_merge = merge_config.get("tasks", "all")
 input_ids_to_merge = merge_config.get("inputs", "all")
 
@@ -559,24 +560,27 @@ if _template_name != "tcb":
     if os.path.exists(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile.sdk"):
         cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile.sdk") .
 
-    cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile") .
-    cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/docker-compose.yml") .
+    if os.path.exists(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile"):
+        cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile") .
+        cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/docker-compose.yml") .
     cp -f @(f"{os.environ['HOME']}/.apollox/assets/github/workflows/build-application.yaml") .
     cp -f @(f"{os.environ['HOME']}/.apollox/assets/gitlab/.gitlab-ci.yml") .
 
     # If there is a .dockerignore file, also include it
     if os.path.exists(f"{os.environ['HOME']}/.apollox/{_template_name}/.dockerignore"):
-        cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/.dockerignore") .
+            cp -f @(f"{os.environ['HOME']}/.apollox/{_template_name}/.dockerignore") .
 
     # ----------------------------------------------------------------- TORIZONPACKAGES.JSON
     with open(f"{os.environ['HOME']}/.apollox/assets/json/torizonPackages.json", "r") as f:
         _torPackagesJson = json.load(f)
 
     # Check also the build part of Dockerfile, for the presence of torizon_packages_build
-    with open(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile", "r") as f:
-        dockerfileLines = f.readlines()
+    buildDepDockerfile = None
+    if os.path.exists(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile"):
+        with open(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile", "r") as f:
+            dockerfileLines = f.readlines()
 
-    buildDepDockerfile = any("torizon_packages_build" in line for line in dockerfileLines)
+        buildDepDockerfile = any("torizon_packages_build" in line for line in dockerfileLines)
 
     if os.path.exists(f"{os.environ['HOME']}/.apollox/{_template_name}/Dockerfile.sdk") or buildDepDockerfile:
         _torPackagesJson["buildDeps"] = []
@@ -736,18 +740,20 @@ if _template_name != "tcb":
     # DOCKERFILE
     # all projects must have it (less TCB)
     # FIXME: should we not be more generic here? if there is tcb should be more
-    _open_merge_window(
-        f"{project_folder}/.conf/tmp/Dockerfile",
-        f"{project_folder}/Dockerfile"
-    )
+    if os.path.exists(f"{project_folder}/.conf/tmp/Dockerfile"):
+        _open_merge_window(
+            f"{project_folder}/.conf/tmp/Dockerfile",
+            f"{project_folder}/Dockerfile"
+        )
 
     print("✅ Dockerfile", color=Color.GREEN)
 
     # DOCKER-COMPOSE.YML
-    _open_merge_window(
-        f"{project_folder}/.conf/tmp/docker-compose.yml",
-        f"{project_folder}/docker-compose.yml"
-    )
+    if os.path.exists(f"{project_folder}/.conf/tmp/docker-compose.yml"):
+        _open_merge_window(
+            f"{project_folder}/.conf/tmp/docker-compose.yml",
+            f"{project_folder}/docker-compose.yml"
+        )
 
     print("✅ docker-compose.yml", color=Color.GREEN)
